3rd Completed assignment

+++ Binomial.java
@@ -11,47 +11,46 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) { 
+	public static long binomial1(int n, int k) { 
 		if (k > n)
-        return 0;
+        return 0L;
 
         if (k == 0 || k == n) 
-        return 1;
+        return 1L;
 
         return binomial1(n-1, k) + binomial1(n-1, k-1);	
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		//// This function creates a 2D array, say memo, 
 		//// and then initializes all its elements to -1.
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
-		 int memo[][] = new int[n+1][k+1];
+		 long memo[][] = new long[n+1][k+1];
         for (int i = 0; i < memo.length; i++)
         for (int j = 0; j < memo[0].length; j++)
-        memo[i][j] = -1;
+        memo[i][j] = -1L;
         return binomial(n, k, memo);
 	}
 	
 
-	private static int binomial(int n, int k, int[][] memo) {
-		if (memo[n][k] != -1) {
+	private static long binomial(int n, int k, long[][] memo) {
+		if (memo[n][k] != -1L) {
 			return memo[n][k];
 		}
 		// Base case
 		if ((k > n)) {
-		   	memo[n][k] = 0; 
-		   	return 0;
+		   	memo[n][k] = 0L; 
+		   	return 0L;
 		}
 		// Another base case
 		if (k == 0 || k == n) {
-		   	memo[n][k] = 1; 
-		   	return 1;
+		   	memo[n][k] = 1L; 
+		   	return 1L;
 		}
 		memo[n][k] = binomial(n - 1, k, memo) + binomial(n - 1, k - 1, memo);
 		return memo[n][k];
 	}
 }
-
